+++ lib/blocs/weight_bloc/weight_bloc.dart
@@ -15,7 +15,7 @@ class WeightBloc extends HydratedBloc<WeightEvent, int> {
 
   @override
   int? fromJson(Map<String, dynamic> json) {
-    return json['state'];
+    return json['state'] as int;
   }
 
   @override
